% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxkl.R
\name{coxkl}
\alias{coxkl}
\title{Cox Proportional Hazards Model with KL Divergence}
\usage{
coxkl(z, delta, time, RS, eta_list, tol = 1e-07, Mstop = 50, beta = NULL)
}
\arguments{
\item{z}{A matrix of covariates where rows represent observations and columns represent variables.}

\item{delta}{A vector of time-to-event outcomes, where 1 indicates an event has occurred and 0 indicates censoring.}

\item{time}{A vector of event times corresponding to each observation.}

\item{RS}{A vector or matrix of external risk scores corresponding to each observation, used for data integration.}

\item{eta_list}{A vector of tuning parameters for data integration, allowing for adjustment in the influence of the external risk scores.}

\item{tol}{The tolerance level for the convergence criterion of the optimization algorithm. Default is 1.0e-7.}

\item{Mstop}{The maximum number of iteration steps for the optimization algorithm. Default is 50.}
}
\value{
A list containing three elements:
\itemize{
\item \code{betas}: A list of estimated coefficient matrices, one for each value of eta.
\item \code{LPs_train}: A list of linear predictors for the training set, one for each value of eta.
\item \code{eta}: The original list of eta values provided as input.
}
}
\description{
This function estimates the coefficients of a Cox proportional hazards model using
Kullback-Leibler divergence for data integration, allowing for the incorporation of
an external risk score (RS) and adjustment of the integration via a list of tuning parameters (eta).
}
\examples{
# Load example data (assuming the package is installed and loaded)
data(simulatedData)
# Fit the model with a single eta value
result1 <- coxkl(z = z, delta = delta, time = time,
                 RS = rs_external1, eta_list = c(0,1,2,3,4))

}
