% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxkl_highdim.R
\name{cv.coxkl_highdim}
\alias{cv.coxkl_highdim}
\title{Cox Proportional Hazards Model with KL Divergence}
\usage{
cv.coxkl_highdim(
  z,
  delta,
  time,
  group,
  theta_tilde,
  eta_kl = 0,
  ...,
  nfolds = 10,
  seed,
  fold,
  se = c("quick", "bootstrap"),
  cv.method = c("LinPred", "VVH"),
  returnY = FALSE,
  trace = FALSE
)
}
\description{
This function estimates the coefficients of a Cox proportional hazards model using
Kullback-Leibler divergence for data integration, allowing for the incorporation of
an external risk score (RS) and adjustment of the integration via a list of tuning parameters (eta).
}
