% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.coxklpred}
\alias{plot.coxklpred}
\title{Plot the relationship between performance score vs. \eqn{\eta}}
\usage{
\method{plot}{coxklpred}(
  object,
  line_size = 1,
  point_size = 3,
  methodline_color = "#4682B4",
  baseline_color = "#778899",
  method_linetype = "solid",
  baseline_linetype = "dashed"
)
}
\arguments{
\item{object}{A model fitted from \code{cv.coxkl}.}

\item{line_size}{Size of the lines.}

\item{point_size}{Size of the points.}

\item{methodline_color}{Color of the}

\item{baseline_color}{Color for the baseline point & segment}

\item{method_linetype}{Line type}

\item{baseline_linetype}{Line type}

\item{legend_title}{Title displayed above the color/linetype legend.}

\item{method_name}{Label used for the main method in legend, color, and linetype.}

\item{baseline_name}{Label used for the baseline in legend, color, and linetype.}
}
\value{
A ggplot object representing the relationship between performance score and eta.
}
\description{
This function takes a model fitted from \code{cv.coxkl} to plot the
relationship between performance score and corresponding \eqn{\eta}.
}
\examples{
data(ExampleData)
library(survival)
result <- cv.coxkl(z = ExampleData$z, delta = ExampleData$status, 
                    time = ExampleData$time, beta = ExampleData$beta_external,
                    eta_list = seq(0, 5, 1))
plot <- plot_FS(result)

}
