% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\name{simulatedData}
\alias{simulatedData}
\title{Simulated Data}
\format{
A list with 2 elements:
\describe{
\item{beta_external1}{A vector representing the parameters from Cox model parameters with good quality.}
\item{beta_external2}{A vector representing the parameters from Cox model parameters with good quality.}
\item{data_internal_list}{A list containing internal data elements with 100 replicates.}
}
}
\usage{
data(simulatedData)
}
\description{
This dataset contains simulated data intended for examples
and testing within the package.
}
