% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.coxkl.R
\name{cv.coxkl}
\alias{cv.coxkl}
\title{Cox Proportional Hazards Model with KL Divergence using Cross-Validation to select best tuning parameter.}
\usage{
cv.coxkl(
  z,
  delta,
  time,
  RS,
  beta,
  eta_list,
  tol = 1e-07,
  Mstop = 50,
  nfolds = 5,
  criteria = "C-Index"
)
}
\arguments{
\item{z}{A matrix of covariates where rows represent observations and columns represent variables.}

\item{delta}{A vector of time-to-event outcomes, where 1 indicates an event has occurred and 0 indicates censoring.}

\item{time}{A vector of event times corresponding to each observation.}

\item{RS}{A vector or matrix of external risk scores corresponding to each observation, used for data integration.}

\item{eta_list}{A vector of tuning parameters for data integration, allowing for adjustment in the influence of the external risk scores.}

\item{tol}{The tolerance level for the convergence criterion of the optimization algorithm. Default is 1.0e-7.}

\item{Mstop}{The maximum number of iteration steps for the optimization algorithm. Default is 50.}

\item{nfolds}{The number of cross-validation folds. Default is 5.}

\item{criteria}{The criterion used to evaluate model performance during
cross-validation.}
}
\value{
A list containing:
\item{result}{A vector of performance scores (e.g., C-index) for each \code{eta}
value in \code{eta_list}. The length of this vector is equal to the length
of \code{eta_list}.}
\item{eta}{The \code{eta_list} provided as input.}
\item{criteria}{The criterion used to evaluate model performance during
cross-validation.}
}
\description{
This function estimates the coefficients of a Cox proportional hazards model using
Kullback-Leibler divergence for data integration, allowing for the incorporation of
an external risk score (RS) and adjustment of the integration via a list of tuning parameters (eta).
}
\examples{
data(simulatedData)
cv_result <- cv.coxkl(z = z, delta = delta, time = time,
                       RS = rs_external1, eta_list = eta_list,
                       nfolds = 5, criteria = "C-Index")
                       
}
